requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.2
anyio==3.7.1
apache-airflow==2.6.3
apache-airflow-providers-apache-livy==3.5.2
apache-airflow-providers-apache-spark==4.1.3
apache-airflow-providers-common-sql==1.7.0
apache-airflow-providers-databricks==4.3.3
apache-airflow-providers-ftp==3.5.0
apache-airflow-providers-http==4.5.0
apache-airflow-providers-imap==3.3.0
apache-airflow-providers-microsoft-mssql==3.4.2
apache-airflow-providers-sqlite==3.4.3
apache-airflow-providers-tableau==4.2.1
apispec==5.2.2
apyori==1.1.2
argcomplete==3.1.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
blis==0.7.10
cachelib==0.9.0
cachetools==5.3.1
catalogue==2.0.9
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
clickclick==20.10.2
cmake==3.27.2
colorama==0.4.6
colorlog==4.8.0
confection==0.1.1
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.3
cymem==2.0.7
databricks-api==0.9.0
databricks-cli==0.17.7
databricks-sql-connector==2.9.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.2
findspark==2.0.1
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
frozenlist==1.4.0
future==0.18.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-re2==1.1
googleapis-common-protos==1.60.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==4.13.0
importlib-resources==6.0.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
keras-nightly==2.14.0.dev2023080207
langcodes==3.3.0
lazy-object-proxy==1.9.0
libclang==16.0.6
limits==3.5.0
linkify-it-py==2.0.2
lit==16.0.6
lockfile==0.12.2
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pathspec==0.9.0
pathy==0.10.2
pendulum==2.1.2
Pillow==10.0.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
pluggy==1.2.0
preshed==3.0.8
prison==0.2.1
protobuf==4.24.0
psutil==5.9.5
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyHive==0.6.5
PyJWT==2.8.0
pymssql==2.2.8
pyodbc==4.0.39
PySocks==1.7.1
pyspark==3.4.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.5.2
rich-argparse==1.2.0
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
selenium==4.11.2
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
srsly==2.4.7
sympy==1.12
tableauserverclient==0.25
tabulate==0.9.0
tb-nightly==2.14.0a20230808
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
text-unidecode==1.3
tf-estimator-nightly==2.14.0.dev2023080308
tf-nightly==2.14.0.dev20230706
thinc==8.1.12
threadpoolctl==3.2.0
thrift==0.16.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
undetected-chromedriver==3.5.2
unicodecsv==0.14.1
Unidecode==1.3.6
urllib3==1.26.16
wasabi==1.1.2
webdriver-manager==4.0.0
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
xgboost==1.7.6
yarl==1.9.2
zipp==3.16.2
